test: update test environment handling and add main test for Paymaster integration
